refactor export module by responsibility
