+++ crates/perfgate/src/app/export.rs
@@ -54,162 +54,21 @@
 //! assert!(csv.contains("bench"));
 //! ```
 
-use std::fmt::Write;
+mod escape;
+mod format;
+mod formatters;
+mod rows;
 
-use perfgate_types::{CompareReceipt, Metric, MetricStatus, RunReceipt};
+pub use escape::csv_escape;
+pub use format::ExportFormat;
+pub use rows::{CompareExportRow, RunExportRow};
 
-/// Supported export formats.
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::app::export::ExportFormat;
-///
-/// let fmt = ExportFormat::Csv;
-/// assert_eq!(ExportFormat::parse("csv"), Some(fmt));
-/// assert_eq!(ExportFormat::parse("unknown"), None);
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ExportFormat {
-    /// RFC 4180 compliant CSV with header row.
-    Csv,
-    /// JSON Lines format (one JSON object per line).
-    Jsonl,
-    /// HTML summary table.
-    Html,
-    /// Prometheus text exposition format.
-    Prometheus,
-    /// JUnit XML format (for legacy CI/Jenkins).
-    JUnit,
-}
+#[cfg(test)]
+pub(crate) use escape::{html_escape, prometheus_escape_label_value};
 
-impl ExportFormat {
-    /// Parse format from string.
-    ///
-    /// ```
-    /// use perfgate::app::export::ExportFormat;
-    ///
-    /// assert_eq!(ExportFormat::parse("csv"), Some(ExportFormat::Csv));
-    /// assert_eq!(ExportFormat::parse("jsonl"), Some(ExportFormat::Jsonl));
-    /// assert_eq!(ExportFormat::parse("prometheus"), Some(ExportFormat::Prometheus));
-    /// assert_eq!(ExportFormat::parse("unknown"), None);
-    /// ```
-    pub fn parse(s: &str) -> Option<Self> {
-        s.parse().ok()
-    }
-}
+use perfgate_types::{CompareReceipt, RunReceipt};
 
-impl std::str::FromStr for ExportFormat {
-    type Err = ();
-
-    /// Parse an [`ExportFormat`] from a string.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use perfgate::app::export::ExportFormat;
-    ///
-    /// let fmt: ExportFormat = "junit".parse().unwrap();
-    /// assert_eq!(fmt, ExportFormat::JUnit);
-    ///
-    /// let bad: Result<ExportFormat, _> = "nope".parse();
-    /// assert!(bad.is_err());
-    /// ```
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "csv" => Ok(ExportFormat::Csv),
-            "jsonl" => Ok(ExportFormat::Jsonl),
-            "html" => Ok(ExportFormat::Html),
-            "prometheus" | "prom" => Ok(ExportFormat::Prometheus),
-            "junit" | "xml" => Ok(ExportFormat::JUnit),
-            _ => Err(()),
-        }
-    }
-}
-
-/// Row structure for RunReceipt export.
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::app::export::RunExportRow;
-///
-/// let row = RunExportRow {
-///     bench_name: "my-bench".into(),
-///     wall_ms_median: 42,
-///     wall_ms_min: 40,
-///     wall_ms_max: 44,
-///     binary_bytes_median: None,
-///     cpu_ms_median: Some(20),
-///     ctx_switches_median: None,
-///     max_rss_kb_median: None,
-///     energy_uj_median: None,
-///     page_faults_median: None,
-///     io_read_bytes_median: None,
-///     io_write_bytes_median: None,
-///     network_packets_median: None,
-///     throughput_median: None,
-///     sample_count: 5,
-///     timestamp: "2024-01-01T00:00:00Z".into(),
-/// };
-/// assert_eq!(row.bench_name, "my-bench");
-/// assert_eq!(row.sample_count, 5);
-/// ```
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct RunExportRow {
-    pub bench_name: String,
-    pub wall_ms_median: u64,
-    pub wall_ms_min: u64,
-    pub wall_ms_max: u64,
-    pub binary_bytes_median: Option<u64>,
-    pub cpu_ms_median: Option<u64>,
-    pub ctx_switches_median: Option<u64>,
-    pub energy_uj_median: Option<u64>,
-    pub max_rss_kb_median: Option<u64>,
-    pub page_faults_median: Option<u64>,
-    pub io_read_bytes_median: Option<u64>,
-    pub io_write_bytes_median: Option<u64>,
-    pub network_packets_median: Option<u64>,
-    pub throughput_median: Option<f64>,
-    pub sample_count: usize,
-    pub timestamp: String,
-}
-
-/// Row structure for CompareReceipt export.
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::app::export::CompareExportRow;
-///
-/// let row = CompareExportRow {
-///     bench_name: "my-bench".to_string(),
-///     metric: "wall_ms".to_string(),
-///     baseline_value: 100.0,
-///     current_value: 110.0,
-///     regression_pct: 10.0,
-///     status: "pass".to_string(),
-///     threshold: 20.0,
-///     warn_threshold: Some(18.0),
-///     cv: None,
-///     noise_threshold: None,
-/// };
-/// assert_eq!(row.metric, "wall_ms");
-/// assert_eq!(row.status, "pass");
-/// ```
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct CompareExportRow {
-    pub bench_name: String,
-    pub metric: String,
-    pub baseline_value: f64,
-    pub current_value: f64,
-    pub regression_pct: f64,
-    pub status: String,
-    pub threshold: f64,
-    pub warn_threshold: Option<f64>,
-    pub cv: Option<f64>,
-    pub noise_threshold: Option<f64>,
-}
+use self::{formatters as out, rows as row_model};
 
 /// Use case for exporting receipts to different formats.
 pub struct ExportUseCase;
@@ -218,7 +77,6 @@ impl ExportUseCase {
     /// Export a [`RunReceipt`] to the specified format.
     ///
     /// ```
-    /// # use std::collections::BTreeMap;
     /// # use perfgate_types::*;
     /// # use perfgate::app::export::{ExportFormat, ExportUseCase};
     /// let receipt = RunReceipt {
@@ -254,36 +112,15 @@ impl ExportUseCase {
     /// assert!(csv.contains("42"));
     /// ```
     pub fn export_run(receipt: &RunReceipt, format: ExportFormat) -> anyhow::Result<String> {
-        let row = Self::run_to_row(receipt);
+        let row = row_model::run_to_row(receipt);
 
         match format {
-            ExportFormat::Csv => Self::run_row_to_csv(&row),
-            ExportFormat::Jsonl => Self::run_row_to_jsonl(&row),
-            ExportFormat::Html => Self::run_row_to_html(&row),
-            ExportFormat::Prometheus => Self::run_row_to_prometheus(&row),
-            ExportFormat::JUnit => Self::run_row_to_junit_run(receipt, &row),
+            ExportFormat::Csv => out::run_row_to_csv(&row),
+            ExportFormat::Jsonl => out::run_row_to_jsonl(&row),
+            ExportFormat::Html => out::run_row_to_html(&row),
+            ExportFormat::Prometheus => out::run_row_to_prometheus(&row),
+            ExportFormat::JUnit => out::run_row_to_junit(receipt, &row),
         }
-    }
-
-    fn run_row_to_junit_run(receipt: &RunReceipt, _row: &RunExportRow) -> anyhow::Result<String> {
-        let mut out = String::new();
-        out.push_str("<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n");
-        out.push_str("<testsuites name=\"perfgate\">\n");
-        writeln!(
-            out,
-            "  <testsuite name=\"{}\" tests=\"1\" failures=\"0\" errors=\"0\">",
-            html_escape(&receipt.bench.name)
-        )?;
-        writeln!(
-            out,
-            "    <testcase name=\"execution\" classname=\"perfgate.{}\" time=\"{}\">",
-            html_escape(&receipt.bench.name),
-            receipt.stats.wall_ms.median as f64 / 1000.0
-        )?;
-        out.push_str("    </testcase>\n");
-        out.push_str("  </testsuite>\n");
-        out.push_str("</testsuites>\n");
-        Ok(out)
     }
 
     /// Export a [`CompareReceipt`] to the specified format.
@@ -322,482 +159,16 @@ impl ExportUseCase {
         receipt: &CompareReceipt,
         format: ExportFormat,
     ) -> anyhow::Result<String> {
-        let rows = Self::compare_to_rows(receipt);
+        let rows = row_model::compare_to_rows(receipt);
 
         match format {
-            ExportFormat::Csv => Self::compare_rows_to_csv(&rows),
-            ExportFormat::Jsonl => Self::compare_rows_to_jsonl(&rows),
-            ExportFormat::Html => Self::compare_rows_to_html(&rows),
-            ExportFormat::Prometheus => Self::compare_rows_to_prometheus(&rows),
-            ExportFormat::JUnit => Self::compare_rows_to_junit(receipt, &rows),
+            ExportFormat::Csv => out::compare_rows_to_csv(&rows),
+            ExportFormat::Jsonl => out::compare_rows_to_jsonl(&rows),
+            ExportFormat::Html => out::compare_rows_to_html(&rows),
+            ExportFormat::Prometheus => out::compare_rows_to_prometheus(&rows),
+            ExportFormat::JUnit => out::compare_rows_to_junit(receipt, &rows),
         }
     }
-
-    /// Convert RunReceipt to exportable row.
-    fn run_to_row(receipt: &RunReceipt) -> RunExportRow {
-        let sample_count = receipt.samples.iter().filter(|s| !s.warmup).count();
-
-        RunExportRow {
-            bench_name: receipt.bench.name.clone(),
-            wall_ms_median: receipt.stats.wall_ms.median,
-            wall_ms_min: receipt.stats.wall_ms.min,
-            wall_ms_max: receipt.stats.wall_ms.max,
-            binary_bytes_median: receipt.stats.binary_bytes.as_ref().map(|s| s.median),
-            cpu_ms_median: receipt.stats.cpu_ms.as_ref().map(|s| s.median),
-            ctx_switches_median: receipt.stats.ctx_switches.as_ref().map(|s| s.median),
-            energy_uj_median: receipt.stats.energy_uj.as_ref().map(|s| s.median),
-            max_rss_kb_median: receipt.stats.max_rss_kb.as_ref().map(|s| s.median),
-            page_faults_median: receipt.stats.page_faults.as_ref().map(|s| s.median),
-            io_read_bytes_median: receipt.stats.io_read_bytes.as_ref().map(|s| s.median),
-            io_write_bytes_median: receipt.stats.io_write_bytes.as_ref().map(|s| s.median),
-            network_packets_median: receipt.stats.network_packets.as_ref().map(|s| s.median),
-            throughput_median: receipt.stats.throughput_per_s.as_ref().map(|s| s.median),
-            sample_count,
-            timestamp: receipt.run.started_at.clone(),
-        }
-    }
-
-    /// Convert CompareReceipt to exportable rows (one per metric, sorted by metric name).
-    fn compare_to_rows(receipt: &CompareReceipt) -> Vec<CompareExportRow> {
-        let mut rows: Vec<CompareExportRow> = receipt
-            .deltas
-            .iter()
-            .map(|(metric, delta)| {
-                let budget = receipt.budgets.get(metric);
-                let threshold = budget.map(|b| b.threshold).unwrap_or(0.0);
-                let warn_threshold = budget.map(|b| b.warn_threshold);
-
-                CompareExportRow {
-                    bench_name: receipt.bench.name.clone(),
-                    metric: metric_to_string(*metric),
-                    baseline_value: delta.baseline,
-                    current_value: delta.current,
-                    regression_pct: delta.pct * 100.0,
-                    status: status_to_string(delta.status),
-                    threshold: threshold * 100.0,
-                    warn_threshold: warn_threshold.map(|t| t * 100.0),
-                    cv: delta.cv.map(|cv| cv * 100.0),
-                    noise_threshold: delta.noise_threshold.map(|t| t * 100.0),
-                }
-            })
-            .collect();
-
-        rows.sort_by(|a, b| a.metric.cmp(&b.metric));
-        rows
-    }
-
-    fn run_row_to_csv(row: &RunExportRow) -> anyhow::Result<String> {
-        let mut output = String::new();
-
-        output.push_str("bench_name,wall_ms_median,wall_ms_min,wall_ms_max,binary_bytes_median,cpu_ms_median,ctx_switches_median,max_rss_kb_median,page_faults_median,io_read_bytes_median,io_write_bytes_median,network_packets_median,energy_uj_median,throughput_median,sample_count,timestamp\n");
-
-        output.push_str(&csv_escape(&row.bench_name));
-        write!(
-            output,
-            ",{},{},{},",
-            row.wall_ms_median, row.wall_ms_min, row.wall_ms_max
-        )?;
-        write_opt_u64(&mut output, row.binary_bytes_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.cpu_ms_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.ctx_switches_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.max_rss_kb_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.page_faults_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.io_read_bytes_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.io_write_bytes_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.network_packets_median);
-        output.push(',');
-        write_opt_u64(&mut output, row.energy_uj_median);
-        output.push(',');
-        if let Some(v) = row.throughput_median {
-            write!(output, "{:.6}", v)?;
-        }
-        write!(output, ",{},", row.sample_count)?;
-        output.push_str(&csv_escape(&row.timestamp));
-        output.push('\n');
-
-        Ok(output)
-    }
-
-    /// Format RunExportRow as JSONL.
-    fn run_row_to_jsonl(row: &RunExportRow) -> anyhow::Result<String> {
-        let json = serde_json::to_string(row)?;
-        let mut out = json;
-        out.push('\n');
-        Ok(out)
-    }
-
-    /// Format CompareExportRows as CSV (RFC 4180).
-    fn compare_rows_to_csv(rows: &[CompareExportRow]) -> anyhow::Result<String> {
-        let mut output = String::new();
-
-        output.push_str(
-            "bench_name,metric,baseline_value,current_value,regression_pct,status,threshold\n",
-        );
-
-        for row in rows {
-            output.push_str(&csv_escape(&row.bench_name));
-            output.push(',');
-            output.push_str(&csv_escape(&row.metric));
-            write!(
-                output,
-                ",{:.6},{:.6},{:.6},",
-                row.baseline_value, row.current_value, row.regression_pct
-            )?;
-            output.push_str(&csv_escape(&row.status));
-            writeln!(output, ",{:.6}", row.threshold)?;
-        }
-
-        Ok(output)
-    }
-
-    /// Format CompareExportRows as JSONL.
-    fn compare_rows_to_jsonl(rows: &[CompareExportRow]) -> anyhow::Result<String> {
-        let mut output = String::new();
-
-        for row in rows {
-            let json = serde_json::to_string(row)?;
-            writeln!(output, "{}", json)?;
-        }
-
-        Ok(output)
-    }
-
-    fn run_row_to_html(row: &RunExportRow) -> anyhow::Result<String> {
-        let html = format!(
-            "<!doctype html><html><head><meta charset=\"utf-8\"><title>perfgate run export</title></head><body>\
-             <h1>perfgate run export</h1>\
-             <table border=\"1\">\
-             <thead><tr><th>bench_name</th><th>wall_ms_median</th><th>wall_ms_min</th><th>wall_ms_max</th><th>binary_bytes_median</th><th>cpu_ms_median</th><th>ctx_switches_median</th><th>max_rss_kb_median</th><th>page_faults_median</th><th>io_read_bytes_median</th><th>io_write_bytes_median</th><th>network_packets_median</th><th>energy_uj_median</th><th>throughput_median</th><th>sample_count</th><th>timestamp</th></tr></thead>\
-             <tbody><tr><td>{bench}</td><td>{wall_med}</td><td>{wall_min}</td><td>{wall_max}</td><td>{binary}</td><td>{cpu}</td><td>{ctx}</td><td>{rss}</td><td>{pf}</td><td>{io_read}</td><td>{io_write}</td><td>{net}</td><td>{energy}</td><td>{throughput}</td><td>{sample_count}</td><td>{timestamp}</td></tr></tbody>\
-             </table></body></html>\n",
-            bench = html_escape(&row.bench_name),
-            wall_med = row.wall_ms_median,
-            wall_min = row.wall_ms_min,
-            wall_max = row.wall_ms_max,
-            binary = row
-                .binary_bytes_median
-                .map_or(String::new(), |v| v.to_string()),
-            cpu = row.cpu_ms_median.map_or(String::new(), |v| v.to_string()),
-            ctx = row
-                .ctx_switches_median
-                .map_or(String::new(), |v| v.to_string()),
-            rss = row
-                .max_rss_kb_median
-                .map_or(String::new(), |v| v.to_string()),
-            pf = row
-                .page_faults_median
-                .map_or(String::new(), |v| v.to_string()),
-            io_read = row
-                .io_read_bytes_median
-                .map_or(String::new(), |v| v.to_string()),
-            io_write = row
-                .io_write_bytes_median
-                .map_or(String::new(), |v| v.to_string()),
-            net = row
-                .network_packets_median
-                .map_or(String::new(), |v| v.to_string()),
-            energy = row
-                .energy_uj_median
-                .map_or(String::new(), |v| v.to_string()),
-            throughput = row
-                .throughput_median
-                .map_or(String::new(), |v| format!("{:.6}", v)),
-            sample_count = row.sample_count,
-            timestamp = html_escape(&row.timestamp),
-        );
-        Ok(html)
-    }
-
-    fn compare_rows_to_html(rows: &[CompareExportRow]) -> anyhow::Result<String> {
-        let mut out = String::from(
-            "<!doctype html><html><head><meta charset=\"utf-8\"><title>perfgate compare export</title></head><body><h1>perfgate compare export</h1><table border=\"1\"><thead><tr><th>bench_name</th><th>metric</th><th>baseline_value</th><th>current_value</th><th>regression_pct</th><th>status</th><th>threshold</th></tr></thead><tbody>",
-        );
-
-        for row in rows {
-            write!(
-                out,
-                "<tr><td>{}</td><td>{}</td><td>{:.6}</td><td>{:.6}</td><td>{:.6}</td><td>{}</td><td>{:.6}</td></tr>",
-                html_escape(&row.bench_name),
-                html_escape(&row.metric),
-                row.baseline_value,
-                row.current_value,
-                row.regression_pct,
-                html_escape(&row.status),
-                row.threshold
-            )?;
-        }
-
-        out.push_str("</tbody></table></body></html>\n");
-        Ok(out)
-    }
-
-    fn run_row_to_prometheus(row: &RunExportRow) -> anyhow::Result<String> {
-        let bench = prometheus_escape_label_value(&row.bench_name);
-        let mut out = String::new();
-        writeln!(
-            out,
-            "perfgate_run_wall_ms_median{{bench=\"{}\"}} {}",
-            bench, row.wall_ms_median
-        )?;
-        writeln!(
-            out,
-            "perfgate_run_wall_ms_min{{bench=\"{}\"}} {}",
-            bench, row.wall_ms_min
-        )?;
-        writeln!(
-            out,
-            "perfgate_run_wall_ms_max{{bench=\"{}\"}} {}",
-            bench, row.wall_ms_max
-        )?;
-        if let Some(v) = row.binary_bytes_median {
-            writeln!(
-                out,
-                "perfgate_run_binary_bytes_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.cpu_ms_median {
-            writeln!(
-                out,
-                "perfgate_run_cpu_ms_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.ctx_switches_median {
-            writeln!(
-                out,
-                "perfgate_run_ctx_switches_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.max_rss_kb_median {
-            writeln!(
-                out,
-                "perfgate_run_max_rss_kb_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.page_faults_median {
-            writeln!(
-                out,
-                "perfgate_run_page_faults_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.io_read_bytes_median {
-            writeln!(
-                out,
-                "perfgate_run_io_read_bytes_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.io_write_bytes_median {
-            writeln!(
-                out,
-                "perfgate_run_io_write_bytes_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.network_packets_median {
-            writeln!(
-                out,
-                "perfgate_run_network_packets_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.energy_uj_median {
-            writeln!(
-                out,
-                "perfgate_run_energy_uj_median{{bench=\"{}\"}} {}",
-                bench, v
-            )?;
-        }
-        if let Some(v) = row.throughput_median {
-            writeln!(
-                out,
-                "perfgate_run_throughput_per_s_median{{bench=\"{}\"}} {:.6}",
-                bench, v
-            )?;
-        }
-        writeln!(
-            out,
-            "perfgate_run_sample_count{{bench=\"{}\"}} {}",
-            bench, row.sample_count
-        )?;
-        Ok(out)
-    }
-
-    fn compare_rows_to_junit(
-        receipt: &CompareReceipt,
-        rows: &[CompareExportRow],
-    ) -> anyhow::Result<String> {
-        let mut out = String::new();
-        let total = rows.len();
-        let failures = rows.iter().filter(|r| r.status == "fail").count();
-        let errors = rows.iter().filter(|r| r.status == "error").count();
-
-        out.push_str("<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n");
-        writeln!(
-            out,
-            "<testsuites name=\"perfgate\" tests=\"{}\" failures=\"{}\" errors=\"{}\">",
-            total, failures, errors
-        )?;
-
-        writeln!(
-            out,
-            "  <testsuite name=\"{}\" tests=\"{}\" failures=\"{}\" errors=\"{}\">",
-            html_escape(&receipt.bench.name),
-            total,
-            failures,
-            errors
-        )?;
-
-        for row in rows {
-            writeln!(
-                out,
-                "    <testcase name=\"{}\" classname=\"perfgate.{}\" time=\"0.0\">",
-                html_escape(&row.metric),
-                html_escape(&receipt.bench.name)
-            )?;
-
-            if row.status == "fail" {
-                write!(
-                    out,
-                    "      <failure message=\"Performance regression detected for {}\">",
-                    html_escape(&row.metric)
-                )?;
-                write!(
-                    out,
-                    "Metric: {}\nBaseline: {:.6}\nCurrent: {:.6}\nRegression: {:.2}%\nThreshold: {:.2}%",
-                    row.metric,
-                    row.baseline_value,
-                    row.current_value,
-                    row.regression_pct,
-                    row.threshold
-                )?;
-                out.push_str("</failure>\n");
-            } else if row.status == "error" {
-                write!(
-                    out,
-                    "      <error message=\"Error occurred during performance check for {}\">",
-                    html_escape(&row.metric)
-                )?;
-                out.push_str("</error>\n");
-            }
-
-            out.push_str("    </testcase>\n");
-        }
-
-        out.push_str("  </testsuite>\n");
-        out.push_str("</testsuites>\n");
-
-        Ok(out)
-    }
-
-    fn compare_rows_to_prometheus(rows: &[CompareExportRow]) -> anyhow::Result<String> {
-        let mut out = String::new();
-        for row in rows {
-            let bench = prometheus_escape_label_value(&row.bench_name);
-            let metric = prometheus_escape_label_value(&row.metric);
-            writeln!(
-                out,
-                "perfgate_compare_baseline_value{{bench=\"{}\",metric=\"{}\"}} {:.6}",
-                bench, metric, row.baseline_value
-            )?;
-            writeln!(
-                out,
-                "perfgate_compare_current_value{{bench=\"{}\",metric=\"{}\"}} {:.6}",
-                bench, metric, row.current_value
-            )?;
-            writeln!(
-                out,
-                "perfgate_compare_regression_pct{{bench=\"{}\",metric=\"{}\"}} {:.6}",
-                bench, metric, row.regression_pct
-            )?;
-            writeln!(
-                out,
-                "perfgate_compare_threshold_pct{{bench=\"{}\",metric=\"{}\"}} {:.6}",
-                bench, metric, row.threshold
-            )?;
-
-            let status_code = match row.status.as_str() {
-                "pass" => 0,
-                "warn" => 1,
-                "fail" => 2,
-                _ => -1,
-            };
-            writeln!(
-                out,
-                "perfgate_compare_status{{bench=\"{}\",metric=\"{}\",status=\"{}\"}} {}",
-                bench,
-                metric,
-                prometheus_escape_label_value(&row.status),
-                status_code
-            )?;
-        }
-        Ok(out)
-    }
-}
-
-/// Convert Metric enum to snake_case string.
-fn metric_to_string(metric: Metric) -> String {
-    metric.as_str().to_string()
-}
-
-/// Convert MetricStatus enum to lowercase string.
-fn status_to_string(status: MetricStatus) -> String {
-    match status {
-        MetricStatus::Pass => "pass".to_string(),
-        MetricStatus::Warn => "warn".to_string(),
-        MetricStatus::Fail => "fail".to_string(),
-        MetricStatus::Skip => "skip".to_string(),
-    }
-}
-
-/// Write an optional u64 value to a buffer. Writes nothing if `None`.
-fn write_opt_u64(buf: &mut String, val: Option<u64>) {
-    if let Some(v) = val {
-        // write! to a String is infallible, unwrap is safe
-        let _ = write!(buf, "{}", v);
-    }
-}
-
-/// Escape a string for CSV per RFC 4180.
-/// If the string contains comma, double quote, or newline, wrap in quotes and escape quotes.
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::app::export::csv_escape;
-///
-/// assert_eq!(csv_escape("hello"), "hello");
-/// assert_eq!(csv_escape("has,comma"), "\"has,comma\"");
-/// assert_eq!(csv_escape("has\"quote"), "\"has\"\"quote\"");
-/// ```
-pub fn csv_escape(s: &str) -> String {
-    if s.contains(',') || s.contains('"') || s.contains('\n') || s.contains('\r') {
-        format!("\"{}\"", s.replace('"', "\"\""))
-    } else {
-        s.to_string()
-    }
-}
-
-fn html_escape(s: &str) -> String {
-    s.replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-}
-
-fn prometheus_escape_label_value(s: &str) -> String {
-    s.replace('\\', "\\\\").replace('"', "\\\"")
 }
 
 #[cfg(test)]
